SYQ API and some debug

+++ APIs/APIs/Models/Users.cs
@@ -52,6 +52,6 @@ namespace APIs.Models
         public string Address { get; set; }
         public string RegDate { get; set; }
         public string Image { get; set; }
-        public long Income { get; set; }
+        public double Income { get; set; }
     }
 }
